Run some Bazel tests on `rules_java` presubmits
This should help catch (most) issues sooner and hopefully obviate the need to make Bazel PRs with unreleased `rules_java` archives.

This will also come in handy when migrating the rules out of `@builtins`.

For now, this is only for linux. I'll add testing on other platforms as a followup.

PiperOrigin-RevId: 595654881
Change-Id: I71d98e94921996155a3d9c68d89df2c1ab938409