Make jmods available from the jdk BUILD file
When using jlink to generate a runtime, it is necessary to point to the jmods
of a JDK for the target architecture. Making these files available in the
java_runtime makes it easier to do this from Bazel.

Since most uses of java_runtime don't need access to the jmods, and so it
would be wasteful to include them in action inputs, we declare
a separate java_runtime for this purpose.

Fixes #136