Do not enable `use_all_repos` for toolchains extension
The toolchain extension isn't internal to rules_java as users have to `use_extension` it to get a reference to `local_jdk`. They should not see a warning about not importing all the other repos, which comes with `modules.as_extension` due to its use of `use_all_repos`. Instead, just declare the extension as reproducible.