Use @com_google_protobuf instead of @protobuf in rules_java
This is to support WORKSPACE mode. Bazel <=7 requires the repository to be called @com_google_protobuf because of the default value of --protocol_compiler flag. To make the migration smooth we need to have the same requirement in Bazel 8.

Upgrade rules_cc to 0.0.15 because of the same thing.
Patch BUILD.java_tools from Bazel with @com_google_protobuf.
Reorder WORKSPACE, so that the correct version of Protobuf is picked up.

PiperOrigin-RevId: 694411552
Change-Id: I1d5d30bc28becc8fe8575846c54bd5ef2d431aa8