Propagate `CcInfo` for native dependencies
This will, in a follow-up, allow us to create and link `java_{binary,test}#laucher`
as the docs suggest.

Unfortunately, `JavaInfo#transitive_native_libraries` fails to propagate
sufficient information about transitive dependencies of the native
libraries, so linking fails because of missing symbols from transitive
libraries that should be there.