Add java_library outputs to validation outputs
This ensures that compilation actions are performed for libraries that are only used as neverlink dependencies, to ensure that any compilation errors are reported.

The minimize the number of additional validation actions, the validation is done for targets that set neverlink, and their transitive runtime dependencies, so anything that is used as a transitive neverlink dep is covered.

PiperOrigin-RevId: 735509081
Change-Id: I76b30e68c8b37eca165861e893ec595cb3040d6c