Initial support for header compilation direct deps
This adds a new output to header compilation actions which is used as the classpath for downstream header compilations, instead of using the same output for downstream header and regular compilations. Using a separate output allows trimming information that isn't needed by header compilation, but is required for regular compilation, including methods. This results in a smaller cache key for header compilation actions.

PiperOrigin-RevId: 764512852
Change-Id: I4895e47f3418e9cbf602edd11fde38de5f3e32ca