Set exec_compatible_with for the runtime Java toolchain
The bootstrap toolchain is already configured this way, and
we want to express compatibility with the execution
platform (where Bazel is executing build actions), rather than with
the target platform.

This allows for using Bazel's --platforms flag to control the target
platform for a build including Java targets, without having the Java
toolchain switched to a JDK that may be incompatible with the execution
platform.