Add support for an unstripped bootclasspath
Running the bootclasspath through ijar was added in https://github.com/bazelbuild/rules_java/pull/324 and was always enabled. This change lets us selectively disable it. This ensures we don't end up depending on the prebuilt ijar in the non-prebuilt toolchain case. This is also needed to bootstrap bazel where we don't have access to `@remote_java_tools`

PiperOrigin-RevId: 828056087
Change-Id: I38402112139221cea606b64110fb1854795a8113